chore: Retire les autres formulaires feedbacks
